requirements.txt
finvizfinance
requests
flask
flask-cors
pandas
python-dotenv
openai
instructor
pydantic
pymongo